+++ NEWS.md
@@ -1,3 +1,12 @@
+# yatah 1.0.0
+
+* `last_clade()`, `last_rank()` and `all_clades()` are now deprecated in favor 
+of `get_last_clade()`, `get_last_rank()` and `get_all_clades()` (#11).  
+
+* New functions `get_clade()` (#12) and `is_at_least_rank()`.
+
+* New dataset `all_ranks`. 
+
 # yatah 0.2.1
 
 * Use of Roxygen 7.2.1.